Add React lines where needed
Summary:
A bunch of our files use React but don't explicitly include it.
This adds the require lines to all of the necessary files (for good
style and so that codemods work correctly).

Also clean up some lint in svg-image.jsx because it was causing arc lint
to fail.

Test Plan:
Run
```
comm -23 <(grep -l -r 'React' src | sort) <(grep -l -r 'require(.react.)' src | sort)
```
and see that the only result is graphie-classes.jsx which only has React
in a comment.

Reviewers: alpert

Reviewed By: alpert

Differential Revision: https://phabricator.khanacademy.org/D22113