Get rid of all references to jed.js.
khan-exercises has been updated with the most recent version of
i18n.js, which as of D22753 is no longer using jed.js.  In fact,
jed.js is no longer included in the khan-exercises repo.  So we have
to get of refernces to it here!

This simplifies a bit of code, in particular the i18n shim file which
can now go away.

I did not update the apiVersion since this is an internal-only
change.  Plus, the latest api hasn't been used in any product yet, so
it's still safe to change it. :-)

Auditors: alex

Test Plan:
Ran 'make forcesubperseus` and then in webapp I ran 'make serve' and
visited
    http://localhost:8080/math/algebra2/systems_eq_ineq/systems_inequalities_precalc/e/modeling-constraints?lang=es
and saw it render (in Spanish) without error.