Add an API option to provide placeholder elements for widgets and images
Summary:
In the side-by-side and group-by-patterns tab of the Manticore exercise
editor, we'd like to render placeholders for widgets and images because
we're rendering each string from an assessment item in a big table and
it was confusing to see a strings from a widget on their own as well as
inside the widget. This diff add two optional properties to apiOptions,
`imagePlaceholder` and `widgetPlaceholder` which take React elements so
that the placeholder can be customized.

Test Plan: make test

Reviewers: emily, jlfwong

Reviewed By: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D23295