Add horizontal scrolling for long math equations in articles
Summary:
In order to accomodate horizontal scrolling, we need to have a new parser rule for math that is block level. The design-request was to have the math-only blocks shrink down so you can see the whole thing on the screen, and then when you tap on it have it expand to look like this and horizontal scroll. This just does the expanded part as a first step.

NOTE: This doesn't address inline-math that is too long – I'm going to say that's a content issue and if we want to be mobile-friendly I think we'll have to just have to make content that doesn't have long equations inline.

Test Plan:
1. Run `make test`
2. Add long math and scroll it:
{F325383}

Reviewers: alex

Reviewed By: alex

Differential Revision: https://phabricator.khanacademy.org/D25609