Add a checkbox in the hint editor for the 'replace' property on hints
Summary:
In https://phabricator.khanacademy.org/D25481 perseus was updated to
render hints with the `replace` property set to `true` by replacing the
previous hint with the new hint.  This diff exposes that functionality
to content creators so that they can tweak exercises ported from
khan-exercises that use this functionality or use it new or existing
perseus exercises.

{F326688}

Test Plan:
- load JSON containing hints with the `replace` property (use ./danae-single.js adding_decimals_0.5 from content-tools-tools)
- see that the 'Replace previous hint' checkboxes are checked for all of the hints
- uncheck some of the checkboxes
- click 'view renderer'
- verify that the previous hint is not replaced for those hints where the checkbox was unchecked

Reviewers: sergei, emily

Reviewed By: emily

Subscribers: alex

Differential Revision: https://phabricator.khanacademy.org/D25620