Enable us to gather statistics on interaction with widgets.
Summary:
This change is part of the [Particle A/B test](https://docs.google.com/document/d/18DXZ6IqrDEfFWkUnmC89vg3g4J0u3MmiUaywO1nTxAQ/edit#) project to help us
determine what aspects of articles are effective.

With this change, it is possible to pass a `trackInteraction` function into
Perseus as an `apiOption`. Most widgets will track the first interaction
of any sort that the user has with it. The graded sequence and graded group
widgets will track each step through the sequence or each click of the
"Check Answer" button as these are useful signals for us.

Test Plan:
1. Go to [the Perseus article editor view](http://localhost:9000/article.html)
2. Try out the widgets that are interactive and watch the console for the
   interaction message.

Reviewers: alex, eliana

Reviewed By: alex

Differential Revision: https://phabricator.khanacademy.org/D25649