Update the eslint-disable directives.
Summary:
We've changed our ka-lint eslintrc recently, removing some rules we
used to enforce, and getting ready to add new ones -- including a big
one: var -> const/let.

This updates the eslint-disable comments at the top of files to
reflect the new eslintrc.

Similar to D25634, I ran this command to undo existing eslint-disable:
   git grep -l '^/\* eslint-disable ' src | xargs perl -nli -e '$o = 1 if m,^/\* eslint-disable,; print unless $o; $o = 0 if m,\*/,'

I then ran
    ~/khan/devtools/khan-linter/runlint.py > /var/tmp/l
    cat /var/tmp/l | grep ^src/ | perl -nle '/^([^:]*):\d+:\d+: [EW]([^ ]*)/ and print "$1 $2"' | sort -u | perl -ane '$t{$F[0]} .= " $F[1],"; if (eof) {for $k (keys %t) { chop $t{$k}; open(OLD, $k); open(NEW, ">$k.tmp"); print NEW qq{/* TODO(csilvers): fix these lint errors (http://eslint.org/docs/rules): */\n/* eslint-disable$t{$k} */\n/* To fix, remove an entry above, run ka-lint, and fix errors. */\n\n}; while (<OLD>) {print NEW $_}; close(OLD); close(NEW); rename("$k.tmp", "$k") }}'

Test Plan:
I ran
   ka-lint
It still prints a lot of errors due to submodules, and a few due to
webpack-* files that I was afraid to change, but src is now clean!

Reviewers: oliver

Subscribers: jlfwong

Differential Revision: https://phabricator.khanacademy.org/D25651