Use webpack externals to make requiring better.
Summary:
Right now, we have a funky setup where we have fake
node_modules/*/index.js files that just export `window.*` values. This takes
advantage of the `externals` feature of webpack to make webpack do the hard
work, instead of us.

This also means that we should be able to do real `require()` and
`module.exports` of perseus inside of webapp, with it correctly requiring the
things that it needs.

Unfortunately, we can't get rid of the fake node_modules/*/index.js files,
because the testing environment depends on them. :(

Test Plan:
 - `make serve`, see that things still look good.
 - `make subperseus`, and see that D25874 works.

Reviewers: alex

Reviewed By: alex

Subscribers: kevinb, jlfwong, jared

Differential Revision: https://phabricator.khanacademy.org/D25872