Fix typo in acceptable formats.
Summary:
Ticket: https://app.asana.com/0/27216215224639/110735518369080

{F501578}

(The awkward upper bar style problem is a separate bug I'm not tackling
as a part of this ticket)

The acceptable format for the cube root of 2 is `sqrt[3]{2}`, and not
`sqrt[3](2)`. This syntax was added in D18923.

Test Plan:
Disable `useMathQuill` and go to
http://localhost:9000/test.html#content=%7B%22question%22%3A%7B%22content%22%3A%22%5B%5B%E2%98%83%20expression%201%5D%5D%5Cn%5Cn%22%2C%22images%22%3A%7B%7D%2C%22widgets%22%3A%7B%22expression%201%22%3A%7B%22type%22%3A%22expression%22%2C%22alignment%22%3A%22default%22%2C%22static%22%3Afalse%2C%22graded%22%3Atrue%2C%22options%22%3A%7B%22answerForms%22%3A%5B%7B%22value%22%3A%22%5C%5Csqrt%5B3%5D%7B2%7D%22%2C%22form%22%3Afalse%2C%22simplify%22%3Afalse%2C%22considered%22%3A%22correct%22%2C%22key%22%3A0%7D%5D%2C%22buttonSets%22%3A%5B%22basic%22%2C%22prealgebra%22%5D%2C%22functions%22%3A%5B%22f%22%2C%22g%22%2C%22h%22%5D%2C%22times%22%3Afalse%7D%2C%22version%22%3A%7B%22major%22%3A1%2C%22minor%22%3A0%7D%7D%7D%7D%2C%22answerArea%22%3A%7B%22calculator%22%3Afalse%2C%22chi2Table%22%3Afalse%2C%22periodicTable%22%3Afalse%2C%22tTable%22%3Afalse%2C%22zTable%22%3Afalse%7D%2C%22itemDataVersion%22%3A%7B%22major%22%3A0%2C%22minor%22%3A1%7D%2C%22hints%22%3A%5B%5D%7D

Type in `sqrt[3]{2}` and see that it's accepted as an answer and is
rendered. Type in `2^(1/3)` and see that it's also accepted as an
answer (but is rendered in a different form).

Reviewers: lauren, sergei

Reviewed By: sergei

Subscribers: alex, jared

Differential Revision: https://phabricator.khanacademy.org/D26892