Don't put fake `//` comments in a CSS file!
Summary: The only valid comments in CSS are `/* */` comments. My
previous changes to the Makefile added some (fake) `//` comments to the
file, which caused the styles to get messed up.

In particular, it looks like it messed up the `.perseus-sr-only`
selector, so I saw this on the SAT page: {F522344}

This turns the top-of-file comment into a real `/* */` comment.

Test Plan:
 - Visit http://localhost:8080/mission/sat/tasks/5681726336532480 and
   see {F522345} (didja see how I didn't have to `make subperseus`,
   because the perseus subrepo is already inside of webapp?! Didja!?)

Auditors: alex