Configure keypad with additional variables and constants
Summary:
This diff implements the scraping logic used to determine the variables and constants that the
keypad must make available when inputting math into an `Expression` widget. In particular, it
builds on top of KAS to scrape the answer forms for the problem and accummulate all of the
necessary variables and symbols, passing them along as part of the keypad's configuration
parameters.

On Asana @ https://app.asana.com/0/114493820703878/132295112211115.

Test Plan:
- Since the config logic isn't hooked up to the keypad yet, add `console.log(this.props.keypadConfiguration)` to the `render` method of `Expression`.
- Open up an `Expression` widget.
    - Verify that there are no extra symbols to start.
- Add an answer with some variables, like `x` and `y`.
    - Verify that `X` and `Y` appear as extra symbols.
- Add some constants, like `pi` and `e` to the answer.
    - Verify that `PI` and `E` appear as extra symbols, along with `X` and `Y`.
- Exercise some other paths, like including a variable inside a function, or using a variable as a function.
- Add some variables to a different answer form and verify that the extra symbols is the cummulative list over all the answer forms.

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27271