Properly sort variables and constants when threading to keypad
Summary:
The variables and constants are meant to be sorted lexicographically in the keypad. This diff
adds that logic by pre-sorting the symbols before we pass them to the keypad. I've put the
constants (Pi and e) at the end, though it's not clear to me how those should be sorted.

Test Plan:
- Create an Expression widget.
- Add an answer with multiple variables and some constants.
    - Verify that they appear sorted, from clockwise top-left, in the rendered keypad.

{F538517}

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27317