Bump the major version to 7
Summary:
In D27271, I updated KAS to include a new `getConsts` function that we're using to scrape constants
from answers in the new mobile web expression editor. As KAS isn't bundled with Perseus, that makes
the change incompatible for the mobile apps, which bundle KAS themselves. Thus, we need to bump the
major version number so that the mobile apps aren't crashing due to `getConsts` being `undefined`.

Test Plan: `make build`

Reviewers: emily, alex

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27374