Handle absence of `buttonProps` in `Expression`
Summary:
`buttonProps` is not a required prop. As such, it can be `null`, which leads to an `undefined is
not a function` when we call `props.buttonProps.includes`. This diff handles that case by
defaulting to the Advanced Expression keypad if `buttonProps` are not provided.

Test Plan: Cross fingers? This only applies to old widgets that haven't been edited in a while, since `buttonSets` has a `defaultProp` now.

Reviewers: benkomalo

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27391