Put new typography styles behind the manatee flag
Summary:
This diff puts all the new typography styling behind the XOM initiative's Manatee flag. To simplify
the experiment and clean-up process, I've re-added all of the old typography styles.

As `framework-perseus` is applied to the root exercise DOM node outside of Perseus, we'll need to
add this flag in webapp as well, at the render sites. Right now, it would be dificult to move this
class application into Perseus, since the `ItemRenderer` setup is such that it modifies sibling DOM
nodes and doesn't render the markup into its own DOM node.

On Asana @ https://app.asana.com/0/132409062806753/137570731984519.

Test Plan:
- Enable `xomManatee` via local storage (e.g., `localStorage.xomMantee = true` in the console).
- Open up an exercise and an article.
- Add some content.
    - Verify that the new XOM styles are being used.
- Disable `xomManatee`.
    - Verify that the content reverts to the old styles.

Reviewers: david, kevinb

Reviewed By: kevinb

Subscribers: benkomalo, jared

Differential Revision: https://phabricator.khanacademy.org/D27636