Fix transformer widget on mobile
Summary:
This changes the transformer widget to `parseFloat` any input it receives.
This is incomplete, b/c the other widget types would need to do it too.

ASANA: https://app.asana.com/0/137742920528627/68171801773007

A larger-reaching but maybe better solution would be to report the grammar as
"just-numbers" or something, and have the mobile stuff pre-convert to numbers
instead of strings.

Test Plan:
Open 'Geometry > Transformations > Determine Translations', and complete some
problems!

Reviewers: charlie

Reviewed By: charlie

Subscribers: #ios

Differential Revision: https://phabricator.khanacademy.org/D27736