Fix lint config in Perseus
Summary:
A while back we've changed things to use a "global" `ka-lint` in
a developers `PATH` instead of requiring that developers have linters
configured to be in a certain spot. My linter is not at this old spot.

Test Plan: - the lint that happens as part of this arc upload

Reviewers: charlie, emily

Reviewed By: emily

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D27753