Only respect touch offset for touch-based gestures
Summary:
In D27836, we modified our movable point behavior to respect the offset at which the user grabbed
the point, to make the gesture more mobile-friendly. In this diff, I've made that behavior
mobile-only, since it's nice to lock to the cursor on desktop.

Note that the check I've used `('ontouchstart' in window)` is consistent with how we check for
touch-based interactions in other movable classes, though it won't give us feature-flagging for
the XOM Manatee beta and may not be totally consistent with how we do mobile detection elsewhere.

Test Plan:
- Open up some interactive widgets on desktop.
    - Verify that, when dragging points, they lock to your cursor, regardless of how far away your cursor was when the gesture started.
- Open up some interactive widgets on a mobile device.
    - Run through the test plan from D27836.

Reviewers: alex, benkomalo, emily

Reviewed By: emily

Subscribers: jared, david

Differential Revision: https://phabricator.khanacademy.org/D27870