Update Makefile to remove React dependencies from math-input's node_modules
Summary: When using the React found in math-input's node_modules, many errors would be thrown when using Perseus. Now, those dependencies are cleaned out so that Perseus's copy of React is used instead.

Test Plan: + Verified in browser that there is no loss of functionality due to using Perseus's copy of React

Reviewers: kevinb

Reviewed By: kevinb

Subscribers: jared, charlie

Differential Revision: https://phabricator.khanacademy.org/D28016