Pass keypad node to `onFocusChange` callback
Summary:
...to allow clients to account for the keypad's presence, when an input is focused.

On the site, we have plenty of page content below our exercises such that there will always be
enough space to scroll an input out from under the newly visible mobile web keypad. However, in our
mobile apps, there's no such accommodation; as such, we need to dynamically add padding to the page
whenever an input is focused. In order to compute the appropriate amount of padding, we need to
provide the native apps with the relevant DOM node.

On Asana @ https://app.asana.com/0/132409062806753/142235863904716.

Test Plan: Coming in the native integration diff.

Reviewers: benkomalo

Reviewed By: benkomalo

Subscribers: jared

Differential Revision: https://phabricator.khanacademy.org/D28046