Add sat-correct and sat-incorrect classnames in perseus review-mode
Summary:
https://app.asana.com/0/110018685008094/117971682493732

{F543450}

We ran into an issue in which explanation/clue text that contained MathJax formulas would show up in the wrong color (black rather than the intended red or green) in SAT review mode. This was because we specify the font-color for `.MathJax .math` in stylesheets/task-package/tasks.less. I added the appropriate style rules for showing MathJax formulas in red or green, and this diff adds the relevant classes from within perseus.

This is my first time touching perseus so I'm not sure I'm even doing this right. This fix might also be a totally gross/wrong way to go about it. Definitely don't love having this mix of CSS and aphrodite styling going on there, but I wasn't sure of a better alternative.

Test Plan:
Confirm that this fixes the bug! MathJax renders red or green appropriately in review mode
Confirm that color of MathJax formulas are otherwise unaffected

Reviewers: lauren, mattdunnrankin

Reviewed By: mattdunnrankin

Subscribers: kejia, tom, jared

Differential Revision: https://phabricator.khanacademy.org/D28037